+++ Example/Sources/InboxViewController.swift
@@ -24,15 +24,40 @@
 
 import UIKit
 import MessageKit
+import SafariServices
 
 final class InboxViewController: UITableViewController {
 
-    let cells = ["Test", "Settings"]
+    let cells = ["Example", "Settings", "Source Code", "Contributers"]
+    
+    // MARK: - View Life Cycle
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        title = "MessageKit"
+        navigationController?.navigationBar.tintColor = .white
+        navigationController?.navigationBar.barTintColor = UIColor(red: 69/255, green: 193/255, blue: 89/255, alpha: 1)
+        navigationController?.navigationBar.titleTextAttributes = [NSAttributedStringKey.foregroundColor: UIColor.white, NSAttributedStringKey.font : UIFont.systemFont(ofSize: 20, weight: UIFont.Weight.bold) ]
         tableView.register(UITableViewCell.self, forCellReuseIdentifier: "cell")
+        tableView.tableFooterView = UIView()
     }
+    
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        if #available(iOS 11.0, *) {
+            navigationController?.navigationBar.prefersLargeTitles = true
+            navigationController?.navigationBar.largeTitleTextAttributes = [NSAttributedStringKey.foregroundColor: UIColor.white]
+        }
+    }
+    
+    override func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        if #available(iOS 11.0, *) {
+            navigationController?.navigationBar.prefersLargeTitles = false
+        }
+    }
+    
+    // MARK: - UITableViewDataSource
 
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return cells.count
@@ -41,16 +66,27 @@ final class InboxViewController: UITableViewController {
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "cell") ?? UITableViewCell()
         cell.textLabel?.text = cells[indexPath.row]
+        cell.accessoryType = .disclosureIndicator
         return cell
     }
+    
+    // MARK: - UITableViewDelegate
     
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let cell = cells[indexPath.row]
         switch cell {
-        case "Test":
+        case "Example":
             navigationController?.pushViewController(ConversationViewController(), animated: true)
         case "Settings":
             navigationController?.pushViewController(SettingsViewController(), animated: true)
+        case "Source Code":
+            guard let url = URL(string: "https://github.com/MessageKit/MessageKit") else { return }
+            let webViewController = SFSafariViewController(url: url)
+            present(webViewController, animated: true, completion: nil)
+        case "Contributers":
+            guard let url = URL(string: "https://github.com/orgs/MessageKit/teams/contributors/members") else { return }
+            let webViewController = SFSafariViewController(url: url)
+            present(webViewController, animated: true, completion: nil)
         default:
             assertionFailure("You need to impliment the action for this cell: \(cell)")
             return