+++ Sources/Layout/MessagesCollectionViewLayoutAttributes.swift
@@ -24,6 +24,7 @@
 
 import UIKit
 
+/// The layout attributes used by a `MessagesCollectionViewCell` to layout its subviews.
 final class MessagesCollectionViewLayoutAttributes: UICollectionViewLayoutAttributes {
 
     // MARK: - Properties