+++ Sources/Models/Avatar.swift
@@ -24,10 +24,20 @@
 
 import Foundation
 
+/// An object used to group the information to be used by an `AvatarView`.
 public struct Avatar {
     
+    // MARK: - Properties
+    
+    /// The image to be used for an `AvatarView`.
     public let image: UIImage?
+    
+    /// The placeholder initials to be used in the case where no image is provided.
+    ///
+    /// The default value of this property is "?".
     public var initals: String = "?"
+    
+    // MARK: - Initializer
     
     public init(image: UIImage? = nil, initals: String = "?") {
         self.image = image