+++ Sources/Models/AvatarAlignment.swift
@@ -24,12 +24,22 @@
 
 import Foundation
 
+/// An enum representing the verical alignment for an `AvatarView`.
 public enum AvatarAlignment {
 
+    /// Aligns the `AvatarView`'s top edge to the cell's top edge.
     case cellTop
-    case messageTop
-    case messageCenter
-    case messageBottom
+    
+    /// Aligns the `AvatarView`'s bottom edge to the cell's bottom edge.
     case cellBottom
-
+    
+    /// Aligns the `AvatarView`'s top edge to the `MessageContainerView`'s top edge.
+    case messageTop
+    
+    /// Aligns the `AvatarView`'s bottom edge to the `MessageContainerView`s bottom edge.
+    case messageBottom
+    
+    /// Aligns the `AvatarView` center to the `MessageContainerView` center.
+    case messageCenter
+    
 }