+++ Sources/Protocols/MessageInputBarDelegate.swift
@@ -25,14 +25,31 @@
 import Foundation
 import UIKit
 
+/// MessageInputBarDelegate is a protocol that can recieve notifications from the MessageInputBar
 public protocol MessageInputBarDelegate: class {
     
+    /// Called when the default send button has been selected
+    ///
+    /// - Parameters:
+    ///   - inputBar: The MessageInputBar
+    ///   - text: The current text in the MessageInputBar's InputTextView
     func messageInputBar(_ inputBar: MessageInputBar, didPressSendButtonWith text: String)
     
+    /// Called when the instrinsicContentSize of the MessageInputBar has changed. Can be used for adjusting content insets
+    /// on other views to make sure the MessageInputBar does not cover up any other view
+    ///
+    /// - Parameters:
+    ///   - inputBar: The MessageInputBar
+    ///   - size: The new instrinsicContentSize
     func messageInputBar(_ inputBar: MessageInputBar, didChangeIntrinsicContentTo size: CGSize)
     
+    /// Called when the MessageInputBar's InputTextView's text has changed. Useful for adding your own logic without the
+    /// need of assigning a delegate or notification
+    ///
+    /// - Parameters:
+    ///   - inputBar: The MessageInputBar
+    ///   - text: The current text in the MessageInputBar's InputTextView
     func messageInputBar(_ inputBar: MessageInputBar, textViewTextDidChangeTo text: String)
-
 }
 
 public extension MessageInputBarDelegate {
@@ -42,5 +59,4 @@ public extension MessageInputBarDelegate {
     func messageInputBar(_ inputBar: MessageInputBar, didChangeIntrinsicContentTo size: CGSize) {}
     
     func messageInputBar(_ inputBar: MessageInputBar, textViewTextDidChangeTo text: String) {}
-
 }