+++ Tests/Mocks/MockMessagesDataSource.swift
@@ -1,18 +1,18 @@
 /*
  MIT License
- 
+
  Copyright (c) 2017 MessageKit
- 
+
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
  in the Software without restriction, including without limitation the rights
  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  copies of the Software, and to permit persons to whom the Software is
  furnished to do so, subject to the following conditions:
- 
+
  The above copyright notice and this permission notice shall be included in all
  copies or substantial portions of the Software.
- 
+
  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -24,35 +24,22 @@
 
 import Foundation
 
-class TestMessagesViewControllerModel: MessagesViewController, MessagesDisplayDelegate {
-    
-    var messageList: [TestMessage] = []
-    
-    static let sender1 = Sender(id: "1", displayName: "Dan")
-    static let sender2 = Sender(id: "2", displayName: "jobs")
-    
-    let testMessage1 = TestMessage(text: "Hi", sender: sender1, messageId: "asdf")
-    let testMessage2 = TestMessage(text: "sup", sender: sender2, messageId: "dddf")
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        messageList = [testMessage1, testMessage2]
-        self.messagesCollectionView.messagesDataSource = self
-    }
-}
+class MockMessagesDataSource: MessagesDataSource {
 
-// - MARK: MessagesDataSource conformace
-extension TestMessagesViewControllerModel: MessagesDataSource {
-    
+    var messages: [MessageType] = []
+    let senders: [Sender] = [Sender(id: "sender_1", displayName: "Sender 1"),
+                             Sender(id: "sender_2", displayName: "Sender 2")]
+
     func currentSender() -> Sender {
-        return Sender(id: "1", displayName: "Dan")
+        return senders[0]
     }
-    
+
     func numberOfMessages(in messagesCollectionView: MessagesCollectionView) -> Int {
-        return messageList.count
+        return messages.count
     }
-    
+
     func messageForItem(at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> MessageType {
-        return messageList[indexPath.section]
+        return messages[indexPath.section]
     }
+
 }