+++ Tests/ViewsTests/MessagesCollectionViewTests.swift
@@ -1,18 +1,18 @@
 /*
  MIT License
-
+ 
  Copyright (c) 2017 MessageKit
-
+ 
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
  in the Software without restriction, including without limitation the rights
  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  copies of the Software, and to permit persons to whom the Software is
  furnished to do so, subject to the following conditions:
-
+ 
  The above copyright notice and this permission notice shall be included in all
  copies or substantial portions of the Software.
-
+ 
  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,20 +23,28 @@
  */
 
 import XCTest
+@testable import MessageKit
 
-final class MessageKitTests: XCTestCase {
+class MessagesCollectionViewTests: XCTestCase {
+
+    var messagesCollectionView: MessagesCollectionView!
+    let rect = CGRect(x: 0, y: 0, width: 100, height: 100)
+    let layout = MessagesCollectionViewFlowLayout()
 
     override func setUp() {
         super.setUp()
+        messagesCollectionView = MessagesCollectionView(frame: rect, collectionViewLayout: layout)
     }
 
     override func tearDown() {
+        messagesCollectionView = nil
         super.tearDown()
     }
 
-    func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    func testInit() {
+        XCTAssertEqual(messagesCollectionView.frame, rect)
+        XCTAssertEqual(messagesCollectionView.collectionViewLayout, layout)
+        XCTAssertEqual(messagesCollectionView.backgroundColor, .white)
     }
 
 }